Don't split a loop task unless the local deque is empty
This commit introduces a simple but important change to the way loop
tasks are scheduled at runtime. Previously, whenever a worker scheduled
a loop task and subsequently received a steal request, the worker would
try to split without ever checking its local deque for queued work. With
this change, a worker splits only if it can't send any other tasks.