Refactor splitting of loop tasks to provide a simpler interface
The changes introduced by this commit simplify both runtime and user
code. This is definitely a good thing, but not the entire motivation.
Another goal is to be able to split the currently running task while
polling for incoming steal requests in user code. The current
implementation of polling is in fact too simple: steal requests are
forwarded to other workers once the local deque is empty, without paying
attention to the possibility of splitting. Check, for example, BPC in
combination with loop tasks, where this behavior presents an actual
performance problem.