Optimize barrier implementation
Remove an arbitrary steal request, not necessarily worker 0's. The catch
with this optimization is that worker 0 must not decline steal requests
after a barrier AND create tasks afterwards, or a race with the manager
may lead to early termination detection.

Enable with -DOPTIMIZE_BARRIER. This barrier might replace the current
implementation after sufficient testing.