Fix handling of update messages
This commit fixes a problem that surfaced when a manager had scheduled
loop tasks and received steal requests. Handling or forwarding a steal
request that is meant as an update message violates the "one steal
request per worker" rule, and may result in two workers sending tasks to
one thief.