Send messages to let workers know when to leave scheduling
There is a subtle complication in combination with -DOPTIMIZE_BARRIER:
Because TASKING_EXIT() is usually preceded by TASKING_BARRIER(), one
worker may receive two messages in close succession---one that says
run_dummy_task(), the other that says notify_workers(). If the worker
has yet to read the first message, sending a second message will fail,
unless we increase the size of the channel buffer or wait for the worker
to catch up, that is, wait until channel_send() returns true. This
explains the busy loop in async_action.