Support up to MAXSTEAL steal requests per worker
Additionally, -DOPTIMIZE_BARRIER is replaced by a slightly different
approach to termination detection in conjunction with task barriers.
Cancelling a random worker's steal request before continuing execution
can in fact violate the safety of termination detection. Consider this:

Suppose there are four workers, numbered from 0 to 3. Worker 0 is the
master and worker 1 is the manager, which is in charge of termination
detection and communicating this information to worker 0. Assume that
worker 0 cancels worker 2's steal request at the end of a task barrier
and goes on to create a new task. Worker 0 might pick up the task itself
before receiving worker 2's new steal request, in which case it declines
and forwards the steal request. With worker 2 being idle, as soon as
worker 1 receives worker 2's steal request, termination is detected
again, despite worker 0 being busy, potentially creating new tasks.
There is a chance that worker 0 receives its own steal request and sends
an update in time, keeping worker 1 from detecting termination, but this
is not guaranteed---a typical race condition.

The only solution seems to require cancelling worker 0's steal request.